backend/requirements.txt
flask
python-dotenv
webdriver-manager
selenium
